Remove leadimage placeholder on customer request.
The most customers do not have a leadimage next to the news text.
So a regular news site would only contain placeholder images.
The new default behaviour will not display the placeholder.

The standard leadimage scale should not be changed. The image size
restriction should be done with the `$leadimage-max-width` sass variable.
Because in mobileview the image will be displayed over the whole screen width.
And a cropped image will not fill the whole screen.