Implement mopage news trigger behavior.
The news trigger behavior for the news container provides configuration
fields for setting up automatic notification to the mopage API whenever
a news page is published with ftw.publisher (receiver side).

This requires an ftw.publisher setup as well as a configured
collective.taskqueue.
The notification will be triggered when the ftw.publisher.receiver has
received a news update.