Mopage trigger: trigger on any update of the news.
Trigger the mopage endpoint on any update of the news.
This now also includes adding new children (simplelayout blocks).
This is important because when the news itself is created it does not
yet have any content (=simplelayout blocks).
If the data is pulled from the mopage endpoint at this point the news
will not have any content.

This change will result in multiple updates when creating or updating a
news, which will make sure that everything is up to date in the end.
Since we cannot tell for sure in advance how many changes we will have
it is hard to merge them. In order to have a consistent state at the end
we therefore do not merge them at all.