Mopage: track content changes.
Since the trigger triggers when adding the news and for each added
content block separately, we need to update the modification date in the
data endpoint accordingly in order to let the remote system track
changes and update correctly.
Since Plone's modification date does not and should not change when the
publisher adds content blocks, we need to track those modification
separately in our own modification date.