Fix "broken" navigation link when in listing_view
When a user was in the /listing_view of an ftw.news.NewsListingBlock the navigation item shown in the navigation was directing to the /block_view. We do not want to use this view because it leads to a plain html site without any css and so on, where the user might think something is broken.