Fix mopage text too small
I accidentally checked if the fractions length was smaller than 0 instead of 10000 - fraction. This happened during refactoring... Now we have a better check that also ensures that the text won't be smaller than 1k chars (eg. a text that is 99k chars long would have resulted in a 9char long output)